requirements.txt
# requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.6
langchain>=0.0.335
langgraph>=0.0.10
langchain-openai>=0.0.1
langchain-community>=0.0.10
redis>=4.6.0
aiomysql>=0.2.0
pydantic>=2.4.2